requirements.txt
python-binance
binance-connector
